refactoring: replaced constants into config in shared layer.
