+++ src/shared/api/config/endpoints.ts
